third_party/WebKit/LayoutTests/fast/dom/CSSStyleDeclaration/css-properties-case-sensitive-expected.txt

CONSOLE WARNING: 'getComputedStyle(e).cssXx' (except .cssFloat) is deprecated and will be removed in M50, around April 2016. Please use 'getComputedStyle(e).xx' instead.
This test checks that access to the CSS float property via JavaScript properties on DOM elements is case sensitive. The use of .CssFloat is deprecated and will be removed in M50.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


normal cases

PASS element.style.float is 'left'
PASS element.style.Float is undefined.

"css" prefix

PASS element.style.cssFloat is 'left'
PASS element.style.CssFloat is 'left'
PASS element.style.Cssfloat is undefined.
PASS element.style.cssfloat is undefined.
PASS successfullyParsed is true

TEST COMPLETE
